+++ packages/common/src/projects/_internal/ProjectBusinessRules.ts
@@ -33,7 +33,6 @@ export const ProjectPermissions = [
   "hub:project:workspace:collaborators",
   "hub:project:workspace:content",
   "hub:project:workspace:metrics",
-  "hub:project:workspace:metrics-coming-soon",
   "hub:project:manage",
 ] as const;
 
@@ -123,12 +122,7 @@ export const ProjectPermissionPolicies: IPermissionPolicy[] = [
     dependencies: ["hub:project:workspace", "hub:project:edit"],
   },
   {
-    permission: "hub:project:workspace:metrics-coming-soon",
-    dependencies: ["hub:project:workspace", "hub:project:edit"],
-  },
-  {
     permission: "hub:project:workspace:metrics",
-    availability: ["alpha"], // gate to alpha for just now
     dependencies: ["hub:project:workspace", "hub:project:edit"],
   },
   {